+++ app/src/main/java/cn/wthee/pcrtool/data/db/view/skilltype/ChangeWidth.kt
@@ -1,0 +1,8 @@
+package cn.wthee.pcrtool.data.db.view.skilltype
+
+import cn.wthee.pcrtool.R
+import cn.wthee.pcrtool.data.db.view.SkillActionDetail
+import cn.wthee.pcrtool.utils.getString
+
+// 52：改变单位距离
+fun SkillActionDetail.changeWidth() = getString(R.string.skill_action_type_desc_52)
