+++ app/src/main/java/cn/wthee/pcrtool/data/db/view/skilltype/KillMe.kt
@@ -1,0 +1,9 @@
+package cn.wthee.pcrtool.data.db.view.skilltype
+
+import cn.wthee.pcrtool.R
+import cn.wthee.pcrtool.data.db.view.SkillActionDetail
+import cn.wthee.pcrtool.utils.getString
+import cn.wthee.pcrtool.utils.getTarget
+
+// 30：立即死亡
+fun SkillActionDetail.killMe() = getString(R.string.skill_action_type_desc_30, getTarget())
