+++ app/src/main/java/cn/wthee/pcrtool/data/db/view/skilltype/LimitAttack.kt
@@ -1,0 +1,8 @@
+package cn.wthee.pcrtool.data.db.view.skilltype
+
+import cn.wthee.pcrtool.R
+import cn.wthee.pcrtool.data.db.view.SkillActionDetail
+import cn.wthee.pcrtool.utils.getString
+
+// 47：上限伤害
+fun SkillActionDetail.limitAttack() = getString(R.string.skill_action_type_desc_47)
